Complete Overhaul
After bricking the entire program and realizing I broke the main react cardinal rule, I decided I needed to take the day to redo a large majority of the code so that the information only goes one way, and not up down and everywhere like it was before. It looks a lot cleaner, however there are new issues that need addressing.